requirements.txt
appdirs==1.4.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.17.62
botocore==1.20.62
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.9
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==4.1.2
django-appconf==1.0.4
django-colorfield==0.7.2
django-cors-headers==3.11.0
django-imagekit==4.1.0
django-ninja==0.17.0
django-s3-storage==0.13.6
django-settings-export==1.2.1
gevent==21.12.0
greenlet==1.1.3.post0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
mysql-connector-python==8.0.23
packaging==20.3
pep517==0.8.2
pilkit==2.0
Pillow==9.3.0
progress==1.5
protobuf==3.18.3
psycopg2==2.8.6
pydantic==1.9.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
requests==2.26.0
retrying==1.3.3
s3transfer==0.4.2
six==1.14.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.7
webencodings==0.5.1
zope.event==4.5.0
zope.interface==5.5.1
